requirements.txt
numpy
PyOpenGL
PyOpenGL_accelerate
pyglfw
pyassimp
freetype-py